+++ docsig/messages.py
@@ -11,6 +11,7 @@ docsig.messages
 | E108: return statement documented for property
 | E109: cannot determine whether a return statement should exist or not
 | E110: documented parameter not equal to its respective argument
+| E111: return statement documented for class
 """
 # Exxx: Error
 # E1xx: Docstring
@@ -24,6 +25,7 @@ E107 = "E107: parameter appears to be incorrectly documented"
 E108 = "E108: return statement documented for property"
 E109 = "E109: cannot determine whether a return statement should exist or not"
 E110 = "E110: documented parameter not equal to its respective argument"
+E111 = "E111: return statement documented for class"
 
 # Hxxx: Hint
 # H1xx: Docstring
@@ -31,3 +33,4 @@ H101 = "hint: there appears to be a syntax error that could be causing this"
 H102 = "hint: documentation is sufficient as a getter is the value returned"
 H103 = "hint: it is possible a syntax error could be causing this"
 H104 = "hint: function is missing a docstring"
+H105 = "hint: a class does not return a value during instantiation"